Add orig_rep_cases to reconstruction

+++ epi_inference/reconstruction/tests/test_recon_workflows.py
@@ -9,6 +9,7 @@ from pyutilib.misc import Options as Options
 from epi_inference.engine import driver
 from epi_inference.util import compare_csv, compare_json
 
+keepfiles = False
 
 class TestReconstruct():
     @classmethod
@@ -33,10 +34,11 @@ class TestReconstruct():
         outputdf, golddf = compare_json('./output/recon_countydata1_12011.json', './baseline/recon_countydata1_12011.json')
     
         # cleanup the files we created
-        os.remove('./output/recon_countydata1_all.json')
-        os.remove('./output/recon_countydata1_all_meta.yml')
-        os.remove('./output/recon_countydata1_12011.json')
-        os.remove('./output/recon_countydata1_12011_meta.yml')
+        if not keepfiles:
+            os.remove('./output/recon_countydata1_all.json')
+            os.remove('./output/recon_countydata1_all_meta.yml')
+            os.remove('./output/recon_countydata1_12011.json')
+            os.remove('./output/recon_countydata1_12011_meta.yml')
 
     def test_reconstruct_stochastic(self):
         args = Options()
@@ -51,12 +53,13 @@ class TestReconstruct():
         outputdf, golddf = compare_json('./output/recon_stoch_countydata1_12011.json', './baseline/recon_stoch_countydata1_12011.json')
     
         # cleanup the files we created
-        os.remove('./output/recon_stoch_countydata1_all_38479387.json')
-        os.remove('./output/recon_stoch_countydata1_all_38479387_meta.yml')
-        os.remove('./output/recon_stoch_countydata1_all_39847938.json')
-        os.remove('./output/recon_stoch_countydata1_all_39847938_meta.yml')
-        os.remove('./output/recon_stoch_countydata1_12011.json')
-        os.remove('./output/recon_stoch_countydata1_12011_meta.yml')
+        if not keepfiles:
+            os.remove('./output/recon_stoch_countydata1_all_38479387.json')
+            os.remove('./output/recon_stoch_countydata1_all_38479387_meta.yml')
+            os.remove('./output/recon_stoch_countydata1_all_39847938.json')
+            os.remove('./output/recon_stoch_countydata1_all_39847938_meta.yml')
+            os.remove('./output/recon_stoch_countydata1_12011.json')
+            os.remove('./output/recon_stoch_countydata1_12011_meta.yml')
 
     def test_recon_json2csv(self):
         args = Options()
@@ -74,16 +77,17 @@ class TestReconstruct():
         df_flatten = pd.read_csv('./output/recon_stoch_countydata1_all_38479387_flatten.csv')
         df_narrow = pd.read_csv('./output/recon_stoch_countydata1_all_38479387_narrow.csv')
 
-        assert(df_narrow.shape[1] == (df_flatten.shape[1]-7+2))
-        assert(df_narrow.shape[0] == 7*df_flatten.shape[0])
-        assert(df_narrow.size == (7*df_flatten.shape[0]) * (df_flatten.shape[1]-7+2))
+        assert(df_narrow.shape[1] == (df_flatten.shape[1]-8+2))
+        assert(df_narrow.shape[0] == 8*df_flatten.shape[0])
+        assert(df_narrow.size == (8*df_flatten.shape[0]) * (df_flatten.shape[1]-8+2))
 
         # cleanup the files we created
-        os.remove('./output/recon_stoch_countydata1_all_38479387_flatten.csv')
-        os.remove('./output/recon_stoch_countydata1_all_flatten.csv')
-        os.remove('./output/recon_stoch_countydata1_all_38479387_narrow.csv')
-        os.remove('./output/recon_stoch_countydata1_all_narrow.csv')
-
+        if not keepfiles:
+            os.remove('./output/recon_stoch_countydata1_all_38479387_flatten.csv')
+            os.remove('./output/recon_stoch_countydata1_all_flatten.csv')
+            os.remove('./output/recon_stoch_countydata1_all_38479387_narrow.csv')
+            os.remove('./output/recon_stoch_countydata1_all_narrow.csv')
+        
     def test_recon_summary(self):
         args = Options()
         args.block = 'recon_summary'
@@ -95,5 +99,6 @@ class TestReconstruct():
         outputdf, golddf = compare_csv('./output/recon_stoch_countydata1_all_summary.csv', './baseline/recon_stoch_countydata1_all_summary.csv')
     
         # cleanup the files we created
-        os.remove('./output/recon_stoch_countydata1_all_summary.csv')
+        if not keepfiles:
+            os.remove('./output/recon_stoch_countydata1_all_summary.csv')
